+++ examples/next-perplexity-clone/CHANGELOG.md
@@ -1,5 +1,12 @@
 # next-perplexity-clone
 
+## 0.1.2-beta.0
+
+### Patch Changes
+
+- Updated dependencies [07d4a18]
+  - @microfox/ai-router@2.0.1-beta.0
+
 ## 0.1.1
 
 ### Patch Changes